+++ src/icons/sports/index.ts
@@ -1,0 +1,3 @@
+import Dart from '@/src/icons/sports/Dart';
+
+export { Dart };