+++ src/icons/status-notifications/index.ts
@@ -2,5 +2,6 @@
  * This file exports all icons in the "Status & Notifications" category.
  */
 import Info from '@/src/icons/status-notifications/Info';
+import TickCloud from '@/src/icons/status-notifications/TickCloud';
 
-export { Info };
+export { Info, TickCloud };