+++ src/icons/travel-hotel/index.ts
@@ -1,0 +1,3 @@
+import MetalBriefcase from '@/src/icons/travel-hotel/MetalBriefcase';
+
+export { MetalBriefcase };